Fix regressions in predict_fev
